+++ CHANGELOG.rst
@@ -12,6 +12,8 @@ Changelog
   removed.  Users on older version of OpenSSL will need to upgrade.
 * **BACKWARDS INCOMPATIBLE:** Support for Python 3.6 has been removed.
 * Updated the minimum supported Rust version (MSRV) to 1.56.0, from 1.48.0.
+* Added support for the :class:`~cryptography.x509.OCSPAcceptableResponses`
+  OCSP extension.
 
 .. _v40-0-1: